PDF_Query_using_GoogleGenerativeAI/requirements.txt
ipykernel
python-dotenv
PyPDF2 
faiss-cpu 
google-generativeai
streamlit